--- a/flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/io/StreamTaskSourceInput.java
+++ b/flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/io/StreamTaskSourceInput.java
@@ -157,9 +157,4 @@ public class StreamTaskSourceInput<T> implements StreamTaskInput<T>, Checkpointa
     public SourceOperator<T, ?> getOperator() {
         return operator;
     }
-
-    // Configure StreamTaskSourceInput#emitNext to emit at most one record to the given DataOutput.
-    public void disableEmitNextLoop() {
-        operator.disableEmitNextLoop();
-    }
 }