--- a/flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/tasks/MultipleInputStreamTask.java
+++ b/flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/tasks/MultipleInputStreamTask.java
@@ -82,14 +82,6 @@ public class MultipleInputStreamTask<OUT>
         StreamConfig configuration = getConfiguration();
         ClassLoader userClassLoader = getUserCodeClassLoader();
 
-        // This is needed for StreamMultipleInputProcessor#processInput to preserve the existing
-        // behavior of choosing an input every time a record is emitted. This behavior is good for
-        // fairness between input consumption. But it can reduce throughput due to added control
-        // flow cost on the per-record code path.
-        for (StreamTaskSourceInput<?> input : operatorChain.getSourceTaskInputs()) {
-            input.disableEmitNextLoop();
-        }
-
         InputConfig[] inputs = configuration.getInputs(userClassLoader);
 
         WatermarkGauge[] watermarkGauges = new WatermarkGauge[inputs.length];
@@ -211,6 +203,15 @@ public class MultipleInputStreamTask<OUT>
         } else {
             return triggerSourcesCheckpointAsync(metadata, options);
         }
+    }
+
+    // This is needed for StreamMultipleInputProcessor#processInput to preserve the existing
+    // behavior of choosing an input every time a record is emitted. This behavior is good for
+    // fairness between input consumption. But it can reduce throughput due to added control
+    // flow cost on the per-record code path.
+    @Override
+    public CanEmitBatchOfRecordsChecker getCanEmitBatchOfRecords() {
+        return () -> false;
     }
 
     private boolean isSynchronous(SnapshotType snapshotType) {