@1996fanrui 2023-01-16
Hi @pnowojski @lindong28 @fredia , I have updated the PR to fix the bug that [FLINK-30533](https://issues.apache.org/jira/browse/FLINK-30533) doesn't check recordWriter.isAvailable() in the loop of SourceOperator. Please help take a look in your free time, thanks a lot ~

And I refactored the `canEmitBatchOfRecords` in the first commit according to @pnowojski 's [comment](https://issues.apache.org/jira/browse/FLINK-30533?focusedCommentId=17676585&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-17676585). And after the second commit, the performance is recovered in my Mac.