@lindong28 2023-01-17
Thanks @1996fanrui for fixing the bug! The bug fix LGTM.

I am not sure if we should add the `DataOutput#canEmitBatchOfRecords` thought. Prior to this PR, `DataOutput` is responsible for data flow and it only provides API to emit data/watermark/latency. Classes such as `SourceOperator` and `StreamTask` are responsible for control flow (e.g. mailbox, availability, loop). Having `StreamTask` passes `mailboxHasMail` to `SourceOperator` preserves this separation of data flow and control flow.

This PR changes this abstraction by adding `DataOutput#canEmitBatchOfRecords`. Now `DataOutput` is responsible for control flow as well.

It seems cleaner and simpler to still have DataOutput be responsible only for data flow. Can you explain what is the benefit for this refactor?