@1996fanrui 2023-01-17
Hi @lindong28 , thanks a lot for your review and discuss here.

> Could you explain what is the benefit for this refactor?

As I understand, there are two reasons for adding `canEmitBatchOfRecords` to `DataOutput`:
- `DataOutput` is responsible for emitting data/watermark/latency, and `canEmitBatchOfRecords` let caller know whether the `emit` can be called repeatedly, so it's related to the `emit`.
- The `canEmitBatchOfRecords` may be called in any emit place, such as: `SourceOperator` and `NetworkInput`, if we add `canEmitBatchOfRecords` in `DataOutput`, it's very helpful for code reuse. You can see the last commit of #21697 , I have made an improvement of `NetworkInput`. The `DataOutput#canEmitBatchOfRecords` is easy to reuse.

Please correct me if I'm wrong, thanks~ 

cc @pnowojski