@lindong28 2023-01-17
Hi @1996fanrui, thanks for the explanation.

Regarding the 1st point mentioned above, yes `canEmitBatchOfRecords()` is related to emit int the sense that it mangaes the "control flow" of when emit can be called. So it is probably consistent with my previous comment.

Regarding the 2nd point, I guess you are saying the implementation of this "optimization" after the refactor is simpler. IMO the abstraction (or separate of concern) between classes is generally more important than implementation, unless one implementation is considerably much simpler and easier to read (which is probably not the case here).

I guess ultimately it comes down to whether the extra implementation simplicity is more important than keeping the DataOutput focusing only on data flow. I am inclined to keep the existing abstraction. But I am open for the approach in this PR if both you and @pnowojski prefer this.