@1996fanrui 2023-01-17
Hi @lindong28 , thanks for your quick feedback.

I want to check with you the first question: No matter which class canEmitBatchOfRecords is added to, we should use `Supplier<Boolean> canEmitBatchOfRecords` instead of `Supplier<Boolean> mailboxHasMail;` due to `canEmitBatchOfRecords` should add some other checks and it's more general and abstract, right?

If yes, we can focus on the next question: Where is `canEmitBatchOfRecords` added to?

> Regarding the 1st point mentioned above, yes `canEmitBatchOfRecords()` is related to emit int the sense that it mangaes the "control flow" of when emit can be called. So it is probably consistent with my previous comment.
> 
> Regarding the 2nd point, I guess you are saying the implementation of this "optimization" after the refactor is simpler. IMO the abstraction (or separate of concern) between classes is generally more important than implementation, because good abstraction makes code more intuitive and easier to maintain in the long term. I suppose the difference between the complexity of these two implementations is not significant in this case.
> 
> I guess ultimately it comes down to whether the extra implementation simplicity is more important than keeping the DataOutput focusing only on data flow. I am inclined to keep the existing abstraction. 

Sounds make sense. Let's wait for @pnowojski 's feedback.