@lindong28 2023-01-17
Hi @1996fanrui, yes I agree we should use Supplier<Boolean> canEmitBatchOfRecords.

I believe `canEmitBatchOfRecords` can be passed to `SourceOperator` similar to how `mailboxHasMail ` is passed to `SourceOperator`. Can you see if the approach in this commit [1] works?

[1] https://github.com/apache/flink/compare/master...lindong28:flink:use-task-is-available