@1996fanrui 2023-01-17
> Hi @1996fanrui, yes I agree we should use Supplier canEmitBatchOfRecords.
> 
> I believe `canEmitBatchOfRecords` can be passed to `SourceOperator` similar to how `mailboxHasMail ` is passed to `SourceOperator`. Can you see if the approach in this commit [1] works?
> 
> [1] [lindong28@bbd2055](https://github.com/lindong28/flink/commit/bbd2055f7d6a203af69eeb1879718f953b2bfffd)

I created a POC branch[1] , please see the latest two commits. The first commit is similar to your commit. And I found the `isEmitNextLoopDisabled` can be simplify, so I using the `canEmitBatchOfRecords` instead of `isEmitNextLoopDisabled` in the 2nd commit. WDYT?

[1] https://github.com/1996fanrui/flink/commits/30623/canEmitBatchOfRecords-POC