@lindong28 2023-01-17
Thanks @1996fanrui for the POC. The approach in the 2nd commit looks much simpler than using the `isEmitNextLoopDisabled()`.