@lindong28 2023-01-18
@pnowojski In order to address the problem around following this supplier in IDE, how about we add an internal functional interface in `StreamTask`? It could be similar to the existing `SizeSupplier` interface ([link](https://github.com/apache/flink/blob/master/flink-runtime/src/main/java/org/apache/flink/runtime/metrics/groups/TaskIOMetricGroup.java#L294)).

Suppose the issue around the IDE can be addressed like this, is there any other concern with the POC approach described here [1]?

After thinking about `DataOutput#canEmitBatchOfRecords`, I think that it looks strange to have `DataOutput` determine whether `it can continue to emit records`. The source of truth of `canEmitBatchOfRecords()` comes from things like `mailbox`, and it is more intuitive to keep it outside/decoupled from DataOutput.

[1] https://github.com/1996fanrui/flink/commits/30623/canEmitBatchOfRecords-POC